+++ Planet Paper/Assets/Materials/bulletColor.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8692160ab9bbc4254b924b0e6f5619ef
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
