+++ Planet Paper/Assets/Models/Bullet PS.prefab
@@ -9814,7 +9814,7 @@ Transform:
   m_GameObject: {fileID: 5201612757389516690}
   m_LocalRotation: {x: -0.7071068, y: 0, z: 0, w: 0.7071068}
   m_LocalPosition: {x: -233.84035, y: 3.3123326, z: -21.164328}
-  m_LocalScale: {x: 0.15, y: 0.15, z: 0.15}
+  m_LocalScale: {x: 0.125, y: 0.125, z: 0.125}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 5199034149963293359}
@@ -9918,7 +9918,7 @@ Rigidbody:
   m_IsKinematic: 0
   m_Interpolate: 0
   m_Constraints: 0
-  m_CollisionDetection: 0
+  m_CollisionDetection: 2
 --- !u!114 &6211698486816613942
 MonoBehaviour:
   m_ObjectHideFlags: 0
