+++ Planet Paper/Assets/Scripts/Trampoline.cs
@@ -5,11 +5,13 @@ using UnityEngine;
 public class Trampoline : MonoBehaviour
 {
 
+    public AudioSource jumpSound; 
     public float jumpHeight = 30f;
     void OnTriggerEnter(Collider col){
         if(col.gameObject.CompareTag("Player")){
             Debug.Log("hit");
             PlayerMovement2.velocity.y = Mathf.Sqrt(jumpHeight * -2f * PlayerMovement2.gravity);
+            jumpSound.Play();
         }
     }
 }
