+++ Planet Paper/Assets/Scripts/enemyBullet.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 963a56961424a489a8dc8ff33da7357c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
