--- Planet Paper/Assets/Tropic Store Assets/SimpleNaturePack/SimpleNaturePack_2020.3_HDRP_v1.24.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 58f7e79a87c79824597746e8c1bfe400
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
