--- Planet Paper/Assets/Tropic Store Assets/SimpleNaturePack/SimpleNaturePack_2020.3_URP_v1.24.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f58d25b9e504a4f41ab4a71cf87c1845
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
