--- Planet Paper/Assets/enemyBullets.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 2df35395d1f0342b4b2d5bf05a7ed11e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
