+++ Planet Paper/ProjectSettings/TagManager.asset
@@ -19,7 +19,7 @@ TagManager:
   - Ground
   - WhatIsPlayer
   - WhatIsBullet
-  - 
+  - WhatIsPlane
   - 
   - 
   -
